xiug

+++ src/com/caihong/cms/entity/main/Order.java
@@ -26,7 +26,7 @@ public class Order  implements Serializable{
 	//订单金额不足以购买内容
 	public static final Integer PRE_PAY_STATUS_ORDER_AMOUNT_NOT_ENOUGH=3;
 	
-	private java.lang.Long id;
+	private java.lang.Integer id;
 	private CmsUser user;
 	private CmsUser rewardUser;
 	private java.lang.Double amount;
@@ -60,10 +60,10 @@ public class Order  implements Serializable{
 
 	private int prePayStatus;
 	
-	public java.lang.Long getId() {
+	public java.lang.Integer getId() {
 		return id;
 	}
-	public void setId(java.lang.Long id) {
+	public void setId(java.lang.Integer id) {
 		this.id = id;
 	}
 	public CmsUser getUser() {
